feat(gfx_d3d): land r_water + r_workercmds_common into POSIX/Switch
Wire src/gfx_d3d/r_water.cpp (338 lines) and r_workercmds_common.cpp
(65 lines) into both backends — water surface tessellation queue
(R_AddWaterCells etc.) and the shared worker-cmd type registry
(R_UpdateRemainingEffects / R_UpdateNonDependentEffects /
R_UpdateSpotLightEffect / R_UpdateXModelBoundsDelayed /
R_SkinGfxEntityDelayed).

Source fixes for -Werror clean build:
  - Cast InterlockedExchangeAdd second arg to `long` so the
    volatile-long g_waterLock field type-deduces correctly.

Dup removal in posix_backbone_stubs:
  - Comment out 6 R_Update*Effect* / R_*GfxEntityDelayed /
    TRACK_r_water stubs now provided by upstream.

Satellite stubs:
  - Sys_Set/ResetUpdateSpotLightEffectEvent.
  - Sys_Set/ResetUpdateNonDependentEffectsEvent.
  - Sys_WaitUpdateNonDependentEffectsCompleted.

r_workercmds.cpp held back — drags more renderer-frame helpers.

+++ src/gfx_d3d/r_water.cpp
@@ -44,7 +44,7 @@ void __cdecl R_UploadWaterTextureInternal(water_t **data)
         GenerateMipMaps(D3DFMT_L8, waterGlob.pixels, water);
     }
 
-    InterlockedExchangeAdd(&g_waterLock, -1);
+    InterlockedExchangeAdd(&g_waterLock, static_cast<long>(-1));
 }
 
 void __cdecl WaterFrequenciesAtTime(complex_s *H, const water_t *water, float t)
